// src/generated/create-passenger-name-record.ts
export interface components {
    schemas: {
        CreatePassengerNameRecordRequest: {
            /** @description Orchestrated Create Passenger Name Record API request. */
            CreatePassengerNameRecordRQ: {
// ...
                AirTax?: {
// ...
                    POS?: {
                        /** @description The point of Sale details. */
                        Source?: {
                            /** @description The additional POS details. */
                            TPA_Extensions?: {
                                /** @description The additional POS details. */
                                UserInfo: {
                                    /** @description The AAA city. */
                                    AAACity: {
                                        /**
                                         * @description The AAA city code.
                                         * @example DWF
                                         */
                                        Code: string;
                                    };
                                    /** @description The action code. */
                                    Action?: {
                                        /**
                                         * @description The action code.
                                         * @example TXLAX
                                         */
                                        Code: string;
                                    };
                                    /** @description The agent sine. */
                                    AgentSine: {
                                        /**
                                         * @description The agent sine.
                                         * @example AAS
                                         */
                                        Code: string;
                                    };
                                    /** @description The user branch. */
                                    Branch?: {
                                        /**
                                         * @description The user branch.
                                         * @example 2867
                                         */
                                        Code: number;
                                    };
                                    /** @description The partition ID. */
                                    Partition: {
                                        /**
                                         * @description The partition ID.
                                         * @example AA
                                         */
                                        ID: string;
                                    };
                                    /** @description The service name. */
                                    Service: {
                                        /**
                                         * @description The service name.
                                         * @example V2TAX
                                         */
                                        Name: string;
                                    };
                                    /**
                                     * @description The user set address.
                                     * @example 000000
                                     */
                                    SetAddress?: string;
                                    /** @description The user station. */
                                    Station?: {
                                        /**
                                         * @description The user station.
                                         * @example 925
                                         */
                                        Code: number;
                                    };
                                    /** @description The transaction ID. */
                                    Transaction?: {
                                        /**
                                         * @description The transaction ID.
                                         * @example 16783392
                                         */
                                        ID: number;
                                    };
                                };
                            };
                            /**
                             * @description The sale PCC.
                             * @example TM61
                             */
                            PseudoCityCode?: string;
                        };
                    };
                };
// ...
            };
        };
        CreatePassengerNameRecordResponse: {
// ...
            CreatePassengerNameRecordRS: {
                /**
                 * @description The API version.
                 * @example 2.5.0
                 */
                version?: string;
                ApplicationResults?: components["schemas"]["ApplicationResults"];
// ...
            };
        };
// ...
        ProblemInformation: {
            /** @description Lists specific information from the system and/or other sub-systems invoked during processing. */
            SystemSpecificResults?: {
                /** @description Contains the executed Sabre host system command. */
                HostCommand?: {
                    /**
                     * @description An abbreviation for Line (LN), Interchange Address (IA), and Terminal Address (TA). A three-byte terminal address, usually provided as a six-character hexadecimal value.
                     * @example 0358DC
                     */
                    LNIATA?: string;
                    /**
                     * @description Sabre host system command.
                     * @example 120MAYDFWLAX
                     */
                    content?: string;
                };
                /** @description Lists messages returned by the system. */
                Message?: {
                    /**
                     * @description An abbreviated version of the error or warning code in textual format.
                     * @example WARN.SP.PROVIDER_WARNING
                     */
                    code?: string;
                    /**
                     * @description An indication of the source of error or warning when processing the request.
                     * @example PNR has not been updated successfully, see remaining messages for details.
                     */
                    content?: string;
                }[];
                /**
                 * @description An abbreviated version of the error or warning in textual format.
                 * @example ADDITIONAL ALLOWANCES AND/OR DISCOUNTS MAY APPL
                 */
                ShortText?: string;
                /**
                 * @description If present, this attribute may identify an unknown or misspelled tag that caused an error in processing.
                 *
                 *     It is recommended that the Tag attribute use XPath notation to identify the location of a tag in the event that more than one tag of the same name is present in the document.
                 *
                 *     Alternatively, the tag name alone can be used to identify missing data [Type=ReqFieldMissing].
                 * @example Unknown Element
                 */
                Element?: string;
                /**
                 * @description If returned, allows for batch processing and the identification of the record that failed amongst a group of records. May be a concatenation of the ID of the failed transaction with specific record(s) associated to that transaction.
                 * @example Unknown RecordID
                 */
                RecordID?: string;
                /**
                 * Format: uri
                 * @description The address for online documentation of the error.
                 * @example https://developer.sabre.com
                 */
                DocURL?: string;
                /**
                 * Format: iso8601-date-time
                 * @description Transaction timestamp in ISO 8601 format.
                 * @example 2024-02-13T04:16:10.174-06:00
                 */
                timeStamp?: string;
                /**
                 * @description If returned, provides an XML ID reference to the element for which the results apply.
                 * @example XML IDREF
                 */
                reference?: string;
            }[];
            /**
             * @description Indicates the source of error or warning.
             * @example BusinessLogic
             * @enum {string}
             */
            type?: "Transport" | "Validation" | "Application" | "BusinessLogic";
            /**
             * Format: iso8601-date-time
             * @description Transaction timestamp in ISO 8601 format.
             * @example 2024-02-13T04:16:10.174-06:00
             */
            timeStamp?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}